restructure searchers
