+++ scopes/explorer/command-bar/command-bar.ui.runtime.tsx
@@ -6,10 +6,10 @@ import UIAspect, { UIRuntime, UiUI } from '@teambit/ui';
 import { PubsubAspect, PubsubUI } from '@teambit/pubsub';
 import { isBrowser } from '@teambit/ui-foundation.ui.is-browser';
 import { CommandBar } from '@teambit/explorer.ui.command-bar';
-import { CommandSearcher } from './ui/command-searcher';
 import { CommandBarAspect } from './command-bar.aspect';
 import { commandBarCommands } from './command-bar.commands';
-import { SearchProvider, Keybinding, CommandHandler, CommandId } from './types';
+import { CommandSearcher, SearchProvider } from './searchers';
+import { Keybinding, CommandHandler, CommandId } from './types';
 import { DuplicateCommandError } from './duplicate-command-error';
 import { KeyEvent } from './model/key-event';
 import { CommandBarContext } from './ui/command-bar-context';
