+++ scopes/explorer/command-bar/index.ts
@@ -1,7 +1,8 @@
 export { CommandBarAspect, CommandBarAspect as default } from './command-bar.aspect';
 export { commandBarCommands } from './command-bar.commands';
-export { ComponentSearcher } from './ui/component-searcher';
-export type { ComponentResultPlugin } from './ui/component-searcher';
+export { ComponentSearcher } from './searchers';
 export { CommandBarContext } from './ui/command-bar-context';
+export type { ComponentResultPlugin, SearchProvider } from './searchers';
 export type { CommandBarUI, CommandEntry } from './command-bar.ui.runtime';
-export type { CommandHandler, SearchResult, CommandId, Keybinding, SearchProvider } from './types';
+export type { CommandHandler, CommandId, Keybinding } from './types';
+export type { SearchResult } from '@teambit/explorer.ui.command-bar';
