+++ scopes/explorer/command-bar/searchers/command-searcher/command-result.module.scss
