+++ scopes/explorer/command-bar/searchers/command-searcher/index.ts
