+++ scopes/explorer/command-bar/searchers/component-searcher/component-result.module.scss
