+++ scopes/explorer/command-bar/searchers/component-searcher/component-result.tsx
