+++ scopes/explorer/command-bar/searchers/component-searcher/index.ts
