+++ scopes/explorer/command-bar/searchers/index.ts
@@ -1,0 +1,4 @@
+export { CommandSearcher } from './command-searcher';
+export { ComponentSearcher } from './component-searcher';
+export type { ComponentResultPlugin } from './component-searcher';
+export type { SearchProvider } from './search-provider';
