+++ scopes/explorer/command-bar/searchers/search-provider.ts
@@ -1,0 +1,8 @@
+import { SearchResult } from '@teambit/explorer.ui.command-bar';
+
+export interface SearchProvider {
+  /** provide completions for this search term */
+  search(term: string, limit: number): SearchResult[];
+  /** determines what terms are handled by this searcher. */
+  test(term: string): boolean;
+}
